:tada: Moved admin-route above sanitylive to avoid page-reloads
