+++ aksel.nav.no/website/app/(admin)/AdminStudio.tsx
@@ -3,7 +3,7 @@
 import { NextStudio } from "next-sanity/studio";
 import { useTheme } from "next-themes";
 import { StudioThemeColorSchemeKey } from "sanity";
-import workspaceConfig from "../../../sanity/sanity.config";
+import workspaceConfig from "../../sanity/sanity.config";
 
 function AdminStudio() {
   const { theme, setTheme } = useTheme();
